move MarkdownArticle to article article component directory
