--- src/components/ArticlesOverview/index.ts
@@ -1,5 +1,0 @@
-import { ArticleCard } from "components/ArticlesOverview/ArticleCard";
-import { ArticlesGrid } from "components/ArticlesOverview/ArticlesGrid";
-import { ArticlesOverview } from "components/ArticlesOverview/ArticlesOverview";
-
-export { ArticlesOverview, ArticleCard, ArticlesGrid };
