+++ src/features/Articles/Articles.tsx
@@ -1,6 +1,5 @@
 import { Box } from "@mui/material";
-import { ArticlesOverview } from "components/ArticlesOverview";
-import { MarkdownArticle } from "components/MarkdownArticle";
+import { ArticlesOverview, MarkdownArticle } from "components/article";
 import React from "react";
 import { Route, Routes } from "react-router-dom";
